requirements.txt
fpdf
python-dotenv
requests
hikari-lightbulb
